Separated DummyCore to a separate file, to keep the empty Dummy available
